+++ srcs/init.c
@@ -4,6 +4,8 @@
 #include <string.h>
 #include <netdb.h>
 #include <unistd.h>
+#include <stdlib.h>
+#include <stdio.h>
 
 char	*get_ip_addr(struct addrinfo *addr, char *ip)
 {
@@ -16,6 +18,29 @@ char	*get_ip_addr(struct addrinfo *addr, char *ip)
 		error_exit("inet_ntop error");
 	return (ip);
 }
+
+struct addrinfo	*host_to_addrinfo(char const *hostname)
+{
+	int				s;
+	struct addrinfo	*result;
+	struct addrinfo	hints;
+
+	memset(&hints, 0, sizeof(hints));
+	hints.ai_family = AF_INET;
+	hints.ai_socktype = SOCK_RAW;
+	hints.ai_protocol = IPPROTO_ICMP;
+
+	s = getaddrinfo(hostname, NULL, &hints, &result);
+	if (s != 0)
+	{
+		if (s == -2)
+			error_exit("unknown host");
+		fprintf(stderr, "getaddrinfo: %s\n", gai_strerror(s));
+		exit(EXIT_FAILURE);
+	}
+	return (result);
+}
+
 
 void	init(t_ping *ping, t_args *args)
 {
